Logging
Rocket's logging actually leaves quite a lot to be desired out of the box
in particular there's no way of associating a log line with a particular
request. Including the log line which says the request is completed.
This module allows us to improve on that by giving each request a random id
and allowing us to arrange for that id to be included with most log lines
that are logged during the processing of the request.